+++ src/main/java/si/matjazcerkvenik/eventlogger/db/FlushQueueThread.java
@@ -38,6 +38,8 @@ public class FlushQueueThread extends Thread {
                 throw new RuntimeException(e);
             }
 
+            EventQueue.getInstance().processRequestsQueue();
+
             IDataManager iDataManager = DataManagerFactory.getInstance().getClient();
 
             LogFactory.getLogger().info("FlushQueueThread: queue size=" + EventQueue.getInstance().getQueueSize());